+++ rdagent/scenarios/kaggle/experiment/playground-series-s4e8_template/train.py
@@ -79,11 +79,11 @@ for f in DIRNAME.glob("model/model*.py"):
     X_valid_selected = select_m.select(X_valid.copy())
 
     m = import_module_from_path(f.stem, f)
-    model_l.append((m.fit(X_train_selected, y_train, X_valid_selected, y_valid), m.predict))
+    model_l.append((m.fit(X_train_selected, y_train, X_valid_selected, y_valid), m.predict, select_m))
 
 # 4) Evaluate the model on the validation set
 metrics_all = []
-for model, predict_func in model_l:
+for model, predict_func, select_m in model_l:
     X_valid_selected = select_m.select(X_valid.copy())
     y_valid_pred = predict_func(model, X_valid_selected)
     y_valid_pred = (y_valid_pred > 0.5).astype(int)
@@ -96,7 +96,7 @@ min_index = np.argmin(metrics_all)
 pd.Series(data=[metrics_all[min_index]], index=["MCC"]).to_csv("submission_score.csv")
 
 # 6) Make predictions on the test set and save them
-X_test_selected = select_m.select(X_test.copy())
+X_test_selected = model_l[min_index][2].select(X_test.copy())
 y_test_pred = model_l[min_index][1](model_l[min_index][0], X_test_selected)
 y_test_pred = (y_test_pred > 0.5).astype(int)
